# swagger.yaml
openapi: 3.0.3
info:
  title: ChainGuard AI API
  description: |
    REST API for ChainGuard AI security monitoring platform.
    
    ChainGuard AI provides comprehensive security monitoring for Avalanche blockchain networks through:
    - Real-time Transaction Ingestion
    - AI Threat Analysis
    - Intelligent Alerting
    - Complete Management Interface
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    Get your token by logging in at `/auth/login`.
    
  version: 1.0.0
  contact:
    name: ChainGuard AI Support
    email: support@chainguard.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.chainguard.ai
    description: Production server
...
paths:
...
  /subnets:
    get:
      tags:
        - Subnets
      summary: List all subnets
      description: Get a list of all configured subnets with optional filtering
      operationId: listSubnets
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of subnets to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of subnets to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: isActive
          in: query
          description: Filter by active status
          required: false
          schema:
            type: boolean
        - name: monitoringEnabled
          in: query
          description: Filter by monitoring status
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of subnets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubnetListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Subnets
      summary: Create subnet
      description: Create a new subnet for monitoring (admin only)
      operationId: createSubnet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubnetRequest'
      responses:
        '201':
          description: Subnet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubnetResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin only)
        '409':
          description: Subnet with chain ID already exists
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message description
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    CreateSubnetRequest:
      type: object
      required:
        - name
        - chainId
        - rpcUrl
        - websocketUrl
      properties:
        name:
          type: string
          example: Avalanche C-Chain
        chainId:
          type: string
          example: "43114"
        rpcUrl:
          type: string
          format: uri
          example: https://api.avax.network/ext/bc/C/rpc
        websocketUrl:
          type: string
          format: uri
          example: wss://api.avax.network/ext/bc/C/ws
        description:
          type: string
          example: Avalanche Mainnet C-Chain
...
    Subnet:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: Avalanche C-Chain
        chain_id:
          type: string
          example: "43114"
        rpc_url:
          type: string
          example: https://api.avax.network/ext/bc/C/rpc
        websocket_url:
          type: string
          example: wss://api.avax.network/ext/bc/C/ws
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
          example: true
        monitoring_enabled:
          type: boolean
          example: true
        created_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubnetResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Subnet'
    SubnetListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Subnet'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0
        total:
          type: integer
          example: 250